requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
dateparser==1.2.0
python-dotenv==1.0.1
requests==2.32.3
pypdf==5.1.0
faster-whisper==1.0.3
telethon==1.36.0
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
